// src/constants/Colors.d.ts
declare const Colors: {
  primary: string;
  primaryDark: string;
  primaryLight: string;
  secondary: string;
  background: string;
  surface: string;
  text: string;
  textSecondary: string;
  border: string;
  error: string;
  success: string;
  warning: string;
  info: string;
  google: string;
  white: string;
  gradientStart: string;
  gradientEnd: string;
};

export default Colors;